+++ ntex/CHANGES.md
@@ -2,6 +2,8 @@
 
 ## [0.2.0-b.12] - 2021-02-18
 
+* Handle EINTR in server accept loop
+
 * Fix double registation for accept back-pressure
 
 ## [0.2.0-b.11] - 2021-02-02